third_party/WebKit/LayoutTests/http/tests/inspector/persistence/persistence-navigator-expected.txt

Verify that navigator view removes mapped UISourceCodes.


Running: addFileSystem

Running: waitForUISourceCodes
[largeicon-navigator-frame] top
  [largeicon-navigator-domain] 127.0.0.1:8000
    [largeicon-navigator-folder] inspector
      [largeicon-navigator-folder] persistence
        [largeicon-navigator-folder] resources
          [largeicon-navigator-file] foo.js
        [largeicon-navigator-file] persistence-navigator.html
        [largeicon-navigator-file] persistence-test.js
      [largeicon-navigator-file] debugger-test.js
      [largeicon-navigator-file] inspector-test.js
      [largeicon-navigator-file] isolated-filesystem-test.js
[largeicon-navigator-folder] www [dimmed]
  [largeicon-navigator-folder] inspector/persistence/resources [dimmed]
    [largeicon-navigator-file] foo.js

Running: addFileMapping
[largeicon-navigator-frame] top
  [largeicon-navigator-domain] 127.0.0.1:8000
    [largeicon-navigator-folder] inspector
      [largeicon-navigator-folder] persistence
        [largeicon-navigator-folder] resources
          [largeicon-navigator-file-sync, badge-navigator-file-sync] foo.js
        [largeicon-navigator-file] persistence-navigator.html
        [largeicon-navigator-file] persistence-test.js
      [largeicon-navigator-file] debugger-test.js
      [largeicon-navigator-file] inspector-test.js
      [largeicon-navigator-file] isolated-filesystem-test.js
[largeicon-navigator-folder] www
  [largeicon-navigator-folder] inspector/persistence/resources
    [largeicon-navigator-file-sync, badge-navigator-file-sync] foo.js

Running: removeFileMapping
[largeicon-navigator-frame] top
  [largeicon-navigator-domain] 127.0.0.1:8000
    [largeicon-navigator-folder] inspector
      [largeicon-navigator-folder] persistence
        [largeicon-navigator-folder] resources
          [largeicon-navigator-file] foo.js
        [largeicon-navigator-file] persistence-navigator.html
        [largeicon-navigator-file] persistence-test.js
      [largeicon-navigator-file] debugger-test.js
      [largeicon-navigator-file] inspector-test.js
      [largeicon-navigator-file] isolated-filesystem-test.js
[largeicon-navigator-folder] www [dimmed]
  [largeicon-navigator-folder] inspector/persistence/resources [dimmed]
    [largeicon-navigator-file] foo.js
